requirements.txt
pvporcupine>=3.0.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
numpy>=1.20.0
elevenlabs>=1.0.0
openai>=1.0.0

